Added functional that runs body in it's own error zone
