+++ lib/main.dart
@@ -1,20 +1,11 @@
+import 'dart:async';
+import 'dart:developer';
+
+import 'package:coffee_app/src/app.dart';
 import 'package:flutter/material.dart';
 
 void main() {
-  runApp(const MainApp());
-}
-
-class MainApp extends StatelessWidget {
-  const MainApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: Scaffold(
-        body: Center(
-          child: Text('Hello World!'),
-        ),
-      ),
-    );
-  }
+  runZonedGuarded(() => runApp(const CoffeeShopApp()), (error, stack) {
+    log(error.toString(), name: 'App Error', stackTrace: stack);
+  });
 }
